re-arranged project folder structure for increased readability
