--- src/main/java/com/github/axelliljendal/finance_tracker/auth/LoginRequest.java
@@ -1,30 +1,0 @@
-package com.github.axelliljendal.finance_tracker.auth;
-
-public class LoginRequest {
-    private String email;
-    private String password;
-
-    public LoginRequest() {
-    }
-
-    public LoginRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
