+++ src/main/java/com/github/axelliljendal/finance_tracker/jwt/JwtService.java
@@ -1,5 +1,6 @@
 package com.github.axelliljendal.finance_tracker.jwt;
 
+import com.github.axelliljendal.finance_tracker.config.JwtConfig;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
